add react router / restructure project
