+++ client/src/components/Landing/Landing.css
@@ -1,28 +1,28 @@
-.App {
+.Landing {
   text-align: center;
 }
 
-.App-logo {
-  animation: App-logo-spin infinite 20s linear;
+.Landing-logo {
+  animation: Landing-logo-spin infinite 20s linear;
   height: 80px;
 }
 
-.App-header {
+.Landing-header {
   background-color: #222;
   height: 150px;
   padding: 20px;
   color: white;
 }
 
-.App-title {
+.Landing-title {
   font-size: 1.5em;
 }
 
-.App-intro {
+.Landing-intro {
   font-size: large;
 }
 
-@keyframes App-logo-spin {
+@keyframes Landing-logo-spin {
   from { transform: rotate(0deg); }
   to { transform: rotate(360deg); }
 }
